fix to overwriting CDM. rename files to be more like other games. start on sharing canvas.
