+++ src/gameData.js
@@ -13,7 +13,7 @@ import TwentyOneQuestions from './games/twentyOneQuestions/TwentyOneQuestions.js
 import burst_Forth from "./games/Burst_Forth/burstForth.js";
 import fight from "./games/fight/fight.js";
 import uno from './games/uno/uno.js';
-import pictionary from './games/pictionary';
+import pictionary from './games/pictionary/Pictionary.js';
 
 const gameData = {
